Add requirement txt

+++ backend/routes/evaluation_routes.py
@@ -2,7 +2,7 @@ from typing import Any, Tuple
 from fastapi import APIRouter, Depends, HTTPException, Path, Query
 from bson.objectid import ObjectId
 from services.auth_services import get_current_user
-from dtos import evaluationDto, Response
+from dtos import chairpersonDto, evaluationDb, Response, examinerDto, supervisorDto
 from config import configEvaluation, configUser
 
 # Initialize Router
@@ -10,8 +10,18 @@ router = APIRouter()
 evalCollection = configEvaluation()
 userCollection = configUser()
 
+# +PrepareEvaluation()/
+# +ViewEvaluation()/
+# +UpdateSupervisor() 
+# +ViewPostponedEvaluation()/
+# +viewExaminer()/
+# +AddExaminer()/
+# +UpdateExaminer()/
+# +AddChairPerson()/
+# +viewChairPerson()/
+
 @router.post("/", tags=["evaluation"], status_code=201)
-async def add_evaluation(evaluation: evaluationDto, current_user: Tuple[str, Any] = Depends(get_current_user)):
+async def PrepareEvaluation(evaluation: evaluationDb, current_user: Tuple[str, Any] = Depends(get_current_user)):
     username, role = current_user
     if(role != 1):
         raise HTTPException(status_code=403, detail = "Only Office Assistant can add evaluation")
@@ -19,8 +29,50 @@ async def add_evaluation(evaluation: evaluationDto, current_user: Tuple[str, Any
     result = evalCollection.insert_one(evaluation.model_dump(exclude_none=True))
     return {"_id": str(result.inserted_id), "evaluation": evaluation}
 
+@router.put("/supervisor/{evaluationId}", response_model=Response, tags=["evaluation"])
+async def add_update_supervisor(
+    evaluationId: str,
+    supervisordto: supervisorDto,
+    current_user: Tuple[str, Any] = Depends(get_current_user)
+):
+    username, role = current_user
+ 
+    if role != 1:  # Check if the user has permission
+        raise HTTPException(status_code=403, detail="You do not have the permission to add or update examiners.")
+
+    if not ObjectId.is_valid(evaluationId):  # Validate evaluation ID format
+        raise HTTPException(status_code=400, detail="Invalid evaluation ID format.")
+
+    eval = evalCollection.find_one({"_id": ObjectId(evaluationId)})
+    if not eval:
+        raise HTTPException(status_code=404, detail="Evaluation not found.")
+
+    if eval.get('lockStatus') == True: # Check if nomination is locked
+        raise HTTPException(status_code=403, detail="Evaluation has been locked")
+
+    # Validate supervisor and co-supervisor IDs
+    if supervisordto.supervisorId:
+        supervisor = userCollection.find_one({"userName": supervisordto.supervisorId})
+        if not supervisor:
+            raise HTTPException(status_code=400, detail="Invalid supervisor ID.")
+
+    if supervisordto.coSupervisorId:
+        co_supervisor = userCollection.find_one({"userName": supervisordto.coSupervisorId})
+        if not co_supervisor:
+            raise HTTPException(status_code=400, detail="Invalid co-supervisor ID.")
+
+    # Update the evaluation
+    eval["supervisorId"] = supervisordto.supervisorId
+    eval["coSupervisorId"] = supervisordto.coSupervisorId
+
+    result = evalCollection.update_one({"_id": ObjectId(evaluationId)}, {"$set": eval})
+    if result.matched_count == 0:
+        raise HTTPException(status_code=404, detail="Failed to update evaluation.")
+
+    return Response(response="Evaluation updated successfully", viewModel = eval, status=True)
+
 @router.put("/examiner/{evaluationId}", response_model=Response, tags=["evaluation"])
-async def update_examiner(evaluationId: str, evaluation: evaluationDto, current_user: Tuple[str, Any] = Depends(get_current_user)):
+async def AddUpdateExaminer(evaluationId: str, examinerdto: examinerDto, current_user: Tuple[str, Any] = Depends(get_current_user)):
 
     username, role = current_user
 
@@ -30,15 +82,24 @@ async def update_examiner(evaluationId: str, evaluation: evaluationDto, current_
     if not ObjectId.is_valid(evaluationId):
         raise HTTPException(status_code=400, detail="Invalid evaluation ID format")
     
-    evaluation_data = evaluation.model_dump()
-    if(evaluation.supervisorId):
-        supervisorId = userCollection.find_one({"userName": evaluation.supervisorId})
+    eval = evalCollection.find_one({"_id": ObjectId(evaluationId)})
+    if not eval:
+        raise HTTPException(status_code=404, detail="Evaluation not found.")
+    
+    if eval.get('lockStatus') == True: # Check if nomination is locked
+        raise HTTPException(status_code=403, detail="Evaluation has been locked")
+        
+
+    if(eval.get('supervisorId')):
+        supervisorId = userCollection.find_one({"userName": eval.get('supervisorId')})
         supervisorRole = supervisorId.get("userRole")
-    if(evaluation.coSupervisorId):
-        coSupervisorId = userCollection.find_one({"userName": evaluation.coSupervisorId})
+
+    if(eval.get('coSupervisorId')):
+        coSupervisorId = userCollection.find_one({"userName": eval.get('coSupervisorId')})
         coSupervisorRole = coSupervisorId.get("userRole")
-    if(evaluation.examinerId1):
-        examinerId1 = userCollection.find_one({"userName": evaluation.examinerId1})
+
+    if(examinerdto.examinerId1):
+        examinerId1 = userCollection.find_one({"userName": examinerdto.examinerId1})
         examiner1Role = examinerId1.get("userRole")
 
     if(examiner1Role is not None and examiner1Role < 3):
@@ -49,59 +110,64 @@ async def update_examiner(evaluationId: str, evaluation: evaluationDto, current_
     if(coSupervisorRole is not None and examiner1Role < coSupervisorRole):
         raise HTTPException(status_code=400, detail="Examiner 1 must be a professor")
 
+    eval['examinerId1'] = examinerdto.examinerId1
+    eval['examinerId2'] = examinerdto.examinerId2
+    eval['examinerId3'] = examinerdto.examinerId3
 
-    result = evalCollection.update_one({"_id": ObjectId(evaluationId)}, {"$set": evaluation_data})
+    result = evalCollection.update_one({"_id": ObjectId(evaluationId)}, {"$set": eval})
     if result.matched_count == 0:
         raise HTTPException(status_code=404, detail="Evaluation not found")
-    return Response(response="Evaluation updated successfully", viewModel = evaluation, status=True)
+    return Response(response="Evaluation updated successfully", viewModel = eval, status=True)
 
 @router.put("/chairperson/{evaluationId}", response_model=Response, tags=["evaluation"])
-async def update_chairperson(evaluationId: str, evaluation: evaluationDto, current_user: Tuple[str, Any] = Depends(get_current_user)):
+async def AddChairPerson(evaluationId: str, chairpersondto: chairpersonDto, current_user: Tuple[str, Any] = Depends(get_current_user)):
+
+    username, role = current_user
+
     # Validate evaluation ID format
     if not ObjectId.is_valid(evaluationId):
         raise HTTPException(status_code=400, detail="Invalid evaluation ID format")
     
-    evaluation_data = evaluation.model_dump()
+    # Fetch existing evaluation data
+    eval = evalCollection.find_one({"_id": ObjectId(evaluationId)})
+    if not eval:
+        raise HTTPException(status_code=404, detail="Evaluation not found")
 
-    username, role = current_user
+    if eval.get('lockStatus') == True:  # Check if nomination is locked
+        raise HTTPException(status_code=403, detail="Evaluation has been locked")
 
     # Build query to check chairperson session limits
     query = {}
-    if evaluation.semester:
-        query["semester"] = evaluation.semester
-    if evaluation.chairpersonId:
-        query["chairpersonId"] = evaluation.chairpersonId
+    if chairpersondto.semester:
+        query["semester"] = chairpersondto.semester
+    if chairpersondto.chairpersonId:
+        query["chairpersonId"] = chairpersondto.chairpersonId
 
     # Check if chairperson exceeds session limit
     evaluation_count = evalCollection.count_documents(query)
 
     if role != 2 :
-        raise HTTPException(status_code=403, detail="Only chairperson can assign chairperson")
+        raise HTTPException(status_code=403, detail="Only coordinator can assign chairperson")
     if evaluation_count > 4:
         raise HTTPException(status_code=400, detail="Chairperson cannot chair more than 4 sessions per semester")
-    
-    # Fetch existing evaluation data
-    evaluationDb = evalCollection.find_one({"_id": ObjectId(evaluationId)})
-    if not evaluationDb:
-        raise HTTPException(status_code=404, detail="Evaluation not found")
     
     # Supervisor and co-supervisor roles validation
     supervisorRole, coSupervisorRole, chairpersonRole = None, None, None
 
-    if evaluationDb.get("supervisorId"):
-        supervisor = userCollection.find_one({"userName": evaluationDb.get("supervisorId")})
+    if eval.get("supervisorId"):
+        supervisor = userCollection.find_one({"userName": eval.get("supervisorId")})
         if not supervisor:
             raise HTTPException(status_code=400, detail="Invalid supervisor ID")
         supervisorRole = supervisor.get("userRole")
 
-    if evaluationDb.get("coSupervisorId"):
-        coSupervisor = userCollection.find_one({"userName": evaluationDb.get("coSupervisorId")})
+    if eval.get("coSupervisorId"):
+        coSupervisor = userCollection.find_one({"userName": eval.get("coSupervisorId")})
         if not coSupervisor:
             raise HTTPException(status_code=400, detail="Invalid co-supervisor ID")
         coSupervisorRole = coSupervisor.get("userRole")
 
-    if evaluation.chairpersonId:
-        chairperson = userCollection.find_one({"userName": evaluation.chairpersonId})
+    if chairpersondto.chairpersonId:
+        chairperson = userCollection.find_one({"userName": chairpersondto.chairpersonId})
         if not chairperson:
             raise HTTPException(status_code=400, detail="Invalid chairperson ID")
         chairpersonRole = chairperson.get("userRole")
@@ -115,25 +181,34 @@ async def update_chairperson(evaluationId: str, evaluation: evaluationDto, curre
     
     if supervisorRole is not None and chairpersonRole < supervisorRole:
         raise HTTPException(status_code=400, detail="Chairperson must be a professor")
+    
+    eval['chairpersonId'] = chairpersondto.chairpersonId
 
     # Update evaluation
-    result = evalCollection.update_one({"_id": ObjectId(evaluationId)}, {"$set": evaluation_data})
+    result = evalCollection.update_one({"_id": ObjectId(evaluationId)}, {"$set": eval})
     if result.matched_count == 0:
         raise HTTPException(status_code=404, detail="Evaluation not found")
     
-    return Response(response="Evaluation updated successfully", viewModel=evaluation, status=True)
+    return Response(response="Evaluation updated successfully", viewModel=eval, status=True)
 
 
 @router.delete("/{evaluationId}", tags=["evaluation"])
 async def delete_evaluation(evaluationId: str, current_user: Any = Depends(get_current_user)):
+
+    lockStatus = evalCollection.find_one({"_id": ObjectId(evaluationId)})
+
+    if lockStatus.get('lockStatus') == True:  # Check if nomination is locked
+        raise HTTPException(status_code=403, detail="Evaluation has been locked")
+    
     result = evalCollection.delete_one({"_id": ObjectId(evaluationId)})
     if result.deleted_count == 0:
         raise HTTPException(status_code=404, detail="Evaluation not found")
-    return {"message": "Evaluation deleted successfully"}
+    
+    return Response(response="Evaluation deleted successfully", viewModel=None, status=True)
 
 
 @router.get("/", tags=["evaluation"])
-async def get_evaluations(
+async def ViewEvaluation(
     studentId: str = Query(None),  # Optional query parameter
     supervisorId: str = Query(None),
     coSupervisorId: int = Query(None),
@@ -184,4 +259,30 @@ async def get_evaluations(
     return result
 
 
+@router.put("/lock", response_model=Response, tags=["evaluation"])
+async def lock_nomination(
+    evaluationId: str,
+    lock: bool,
+    current_user: Tuple[str, Any] = Depends(get_current_user)
+):
+    username, role = current_user
+ 
+    if role != 2:  # Check if the user has permission
+        raise HTTPException(status_code=403, detail="You do not have the permission to lock nominations.")
 
+    if not ObjectId.is_valid(evaluationId):  # Validate evaluation ID format
+        raise HTTPException(status_code=400, detail="Invalid evaluation ID format.")
+
+    eval = evalCollection.find_one({"_id": ObjectId(evaluationId)})
+    if not eval:
+        raise HTTPException(status_code=404, detail="Evaluation not found.")
+
+
+    # Update the evaluation
+    eval["lockStatus"] = lock
+
+    result = evalCollection.update_one({"_id": ObjectId(evaluationId)}, {"$set": eval})
+    if result.matched_count == 0:
+        raise HTTPException(status_code=404, detail="Failed to update evaluation.")
+
+    return Response(response="Evaluation updated successfully", viewModel = eval, status=True)
